+++ src/containerapp/azext_containerapp/_help.py
@@ -1915,6 +1915,139 @@ helps['containerapp connected-env list'] = """
           az containerapp connected-env list -g MyResourceGroup
 """
 
+helps['containerapp connected-env dapr-component'] = """
+    type: group
+    short-summary: Commands to manage Dapr components for Container Apps connected environments.
+"""
+
+helps['containerapp connected-env dapr-component list'] = """
+    type: command
+    short-summary: List Dapr components for a connected environment.
+    examples:
+    - name: List Dapr components for a connected environment.
+      text: |
+          az containerapp connected-env dapr-component list -g MyResourceGroup --name MyConnectedEnv
+"""
+
+helps['containerapp connected-env dapr-component show'] = """
+    type: command
+    short-summary: Show the details of a Dapr component.
+    examples:
+    - name: Show the details of a Dapr component.
+      text: |
+          az containerapp connected-env dapr-component show -g MyResourceGroup --dapr-component-name MyDaprComponentName --name MyConnectedEnv
+"""
+
+helps['containerapp connected-env dapr-component set'] = """
+    type: command
+    short-summary: Create or update a Dapr component.
+    examples:
+    - name: Create a Dapr component.
+      text: |
+          az containerapp connected-env dapr-component set -g MyResourceGroup --name MyEnv --yaml MyYAMLPath --dapr-component-name MyDaprComponentName
+"""
+
+helps['containerapp connected-env dapr-component remove'] = """
+    type: command
+    short-summary: Remove a Dapr component from a connected environment.
+    examples:
+    - name: Remove a Dapr component from a Container Apps connected environment.
+      text: |
+          az containerapp connected-env dapr-component remove -g MyResourceGroup --dapr-component-name MyDaprComponentName --name MyConnectedEnv
+"""
+
+helps['containerapp connected-env storage'] = """
+    type: group
+    short-summary: Commands to manage storage for the Container Apps connected environment.
+"""
+
+helps['containerapp connected-env storage list'] = """
+    type: command
+    short-summary: List the storages for a connected environment.
+    examples:
+    - name: List the storages for a connected environment.
+      text: |
+          az containerapp connected-env storage list -g MyResourceGroup -n MyConnectedEnv
+"""
+
+helps['containerapp connected-env storage show'] = """
+    type: command
+    short-summary: Show the details of a storage.
+    examples:
+    - name: Show the details of a storage.
+      text: |
+          az containerapp connected-env storage show -g MyResourceGroup --storage-name MyStorageName -n MyConnectedEnv
+"""
+
+helps['containerapp connected-env storage set'] = """
+    type: command
+    short-summary: Create or update a storage.
+    examples:
+    - name: Create a storage.
+      text: |
+          az containerapp connected-env storage set -g MyResourceGroup -n MyEnv --storage-name MyStorageName --access-mode ReadOnly --azure-file-account-key MyAccountKey --azure-file-account-name MyAccountName --azure-file-share-name MyShareName
+"""
+
+helps['containerapp connected-env storage remove'] = """
+    type: command
+    short-summary: Remove a storage from a connected environment.
+    examples:
+    - name: Remove a storage from a Container Apps connected environment.
+      text: |
+          az containerapp connected-env storage remove -g MyResourceGroup --storage-name MyStorageName -n MyConnectedEnv
+"""
+
+# Certificates Commands
+helps['containerapp connected-env certificate'] = """
+    type: group
+    short-summary: Commands to manage certificates for the Container Apps connected environment.
+"""
+
+helps['containerapp connected-env certificate list'] = """
+    type: command
+    short-summary: List certificates for a connected environment.
+    examples:
+    - name: List certificates for a connected environment.
+      text: |
+          az containerapp connected-env certificate list -g MyResourceGroup --name MyConnectedEnv
+    - name: List certificates by certificate id.
+      text: |
+          az containerapp connected-env certificate list -g MyResourceGroup --name MyConnectedEnv --certificate MyCertificateId
+    - name: List certificates by certificate name.
+      text: |
+          az containerapp connected-env certificate list -g MyResourceGroup --name MyConnectedEnv --certificate MyCertificateName
+    - name: List certificates by certificate thumbprint.
+      text: |
+          az containerapp connected-env certificate list -g MyResourceGroup --name MyConnectedEnv --thumbprint MyCertificateThumbprint
+"""
+
+helps['containerapp connected-env certificate upload'] = """
+    type: command
+    short-summary: Add or update a certificate.
+    examples:
+    - name: Add or update a certificate.
+      text: |
+          az containerapp connected-env certificate upload -g MyResourceGroup --name MyConnectedEnv --certificate-file MyFilepath
+    - name: Add or update a certificate with a user-provided certificate name.
+      text: |
+          az containerapp connected-env certificate upload -g MyResourceGroup --name MyConnectedEnv --certificate-file MyFilepath --certificate-name MyCertificateName
+"""
+
+helps['containerapp connected-env certificate delete'] = """
+    type: command
+    short-summary: Delete a certificate from the Container Apps connected environment.
+    examples:
+    - name: Delete a certificate from the Container Apps connected environment by certificate name
+      text: |
+          az containerapp connected-env certificate delete -g MyResourceGroup --name MyConnectedEnv --certificate MyCertificateName
+    - name: Delete a certificate from the Container Apps connected environment by certificate id
+      text: |
+          az containerapp connected-env certificate delete -g MyResourceGroup --name MyConnectedEnv --certificate MyCertificateId
+    - name: Delete a certificate from the Container Apps connected environment by certificate thumbprint
+      text: |
+          az containerapp connected-env certificate delete -g MyResourceGroup --name MyConnectedEnv --thumbprint MyCertificateThumbprint
+"""
+
 # Container Apps Job Commands
 
 helps['containerapp job create'] = """