+++ src/containerapp/HISTORY.rst
@@ -8,6 +8,9 @@ upcoming
 * Add command group 'az containerapp connected-env', support show/list/delete/create connected environment
 * 'az containerapp create': support --source and --repo properties
 * 'az containerapp update': support --source property
+* Add command group 'az containerapp connected-env certificate', support list/upload/delete connectedEnvironments certificate
+* Add command group 'az containerapp connected-env dapr-component', support list/show/set/remove connectedEnvironments daprComponents
+* Add command group 'az containerapp connected-env storage', support list/show/set/remove connectedEnvironments storage
 
 0.3.39
 ++++++